Renamed rflibLogEventSingleMessage to rflibLogEventViewerMessage
