--- rflib/main/default/lwc/rflibLogEventSingleMessage/rflibLogEventSingleMessage.css
@@ -1,79 +1,0 @@
-.json-message {
-    border: 1px solid #dee2e6;
-    border-radius: 0.375rem;
-    overflow: hidden;
-}
-
-.json-header {
-    display: flex;
-    align-items: center;
-    padding: 0.5rem 0.75rem;
-    border-bottom: 1px solid #dee2e6;
-    cursor: pointer;
-    transition: background-color 0.2s ease;
-}
-
-.json-label {
-    font-weight: 600;
-    font-size: 0.875rem;
-    color: #495057;
-    margin-left: 0.5rem;
-    white-space: nowrap;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    display: inline-block;
-    width: 100%;
-}
-
-.json-content {
-    margin: 0;
-    padding: 0.75rem;
-    background-color: inherit;
-    font-family: 'Monaco', 'Menlo', 'Ubuntu Mono', 'Consolas', 'source-code-pro', monospace;
-    font-size: 0.8125rem;
-    line-height: 1.4;
-    color: #212529;
-    white-space: pre-wrap;
-    word-wrap: break-word;
-    border: none;
-    border-radius: 0;
-    overflow-x: auto;
-}
-
-.json-content code {
-    background: none;
-    color: inherit;
-    font-size: inherit;
-    padding: 0;
-}
-
-.json-preview {
-    padding: 0.75rem;
-    font-style: italic;
-    color: #6c757d;
-    font-size: 0.875rem;
-    background-color: #ffffff;
-}
-
-.text-message {
-    padding: 0.25rem 0;
-    line-height: 1.5;
-}
-
-.text-message lightning-formatted-text {
-    white-space: pre-wrap;
-    word-wrap: break-word;
-}
-
-.wrapper {
-    line-height: 1.5;
-    padding: 5px 10px;
-}
-
-.light {
-    background-color: #ffffff;
-}
-
-.dark {
-    background-color: #f7f7f7;
-}
