+++ rflib/main/default/lwc/rflibLogEventViewerMessage/rflibLogEventViewerMessage.html
@@ -22,9 +22,7 @@
         </template>
         <template lwc:if={message.isText}>
             <div class="text-message">
-                <lightning-formatted-text
-                    value={processedMessage}
-                ></lightning-formatted-text>
+                <lightning-formatted-text value={processedMessage}></lightning-formatted-text>
             </div>
         </template>
     </div>
